feat: adds models for video engagement reports
`video_plays` contains a record for each time a learner played a video.
Can be used to generate reports of total plays and unique watchers

`transcript_usage` contains a record for each time a transcript or
closed caption was enabled. Can be used to generate reports for total
transcription/caption usage as well as the number of unique learners who
enabled transcription/closed captioning