feat: add extended course block dimension table
This change introduces another dimension table,
dim_course_blocks_extended, which includes the section and subsection
names to which each block belongs. This is included as a separate model
instead of adding onto dim_course_blocks so as to avoid any unnecessary
joins when a model requires block metadata but does not need the extra
information joined into dim_course_blocks by this new model.